requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
humanfriendly==10.0
idna==3.7
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
pika-stubs==0.1.3
pika==1.3.2
pip==24.2
protobuf==5.27.1
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
pyreadline3==3.4.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4

dill==0.3.8
captcha==0.6.0
pillow==10.4.0
fastapi==0.112.2
uvicorn==0.30.6
starlette==0.38.2
matplotlib==3.9.2
tabulate==0.9.0
sqids==0.5.0